requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
av==12.3.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
build==1.2.1
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.2
coloredlogs==15.0.1
comm==0.2.2
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
datasets==3.0.0
debugpy==1.8.1
decorator==5.1.1
deepface==0.0.92
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.3.1
django-countries==7.6.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-sslify==0.2.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dlib==19.24.99
dnspython==2.6.1
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
elastic-transport==8.13.1
elasticsearch==8.14.0
executing==2.0.1
facenet-pytorch==2.6.0
fastjsonschema==2.20.0
filelock==3.15.3
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.0
gast==0.6.0
gdown==5.2.0
google-crc32c==1.6.0
google-pasta==0.2.0
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.7
ifaddr==0.2.0
incremental==24.7.1
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_proxy==4.2.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_vpython==3.1.8
jupyterlab_widgets==3.0.11
keras==3.4.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
mtcnn==0.1.1
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nlp==0.4.0
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.19.0
openai==1.35.7
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycoin==0.92.20230326
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pyee==12.0.0
Pygments==2.18.0
PyJWT==2.8.0
pylibsrtp==0.10.0
pymongo==4.8.0
pyOpenSSL==24.1.0
pypng==0.20220715.0
pyproject_hooks==1.1.0
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
PySocks==1.7.1
pytest==8.2.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==2.0.7
pyttsx4==3.0.15
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qrcode==7.4.2
qtconsole==5.5.2
QtPy==2.4.1
rave-python==1.4.0
redis==5.0.7
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
resend==2.0.0
retina-face==0.0.17
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.5.2
scipy==1.14.1
Send2Trash==1.8.3
service-identity==24.1.0
setuptools==73.0.1
shiboken6==6.7.2
simpervisor==1.0.0
SimpleCV==1.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.12.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorrt==10.1.0
tensorrt-cu12==10.1.0
tensorrt-cu12-bindings==10.1.0
tensorrt-cu12-libs==10.1.0
termcolor==2.4.0
terminado==0.18.1
tf_keras==2.17.0
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
twilio==9.2.4
Twisted==24.3.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.1
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
vpython==7.6.5
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.2
wheel==0.44.0
whitenoise==6.6.0
widgetsnbextension==4.0.11
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zope.interface==6.4.post2
